requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
